Update tournament data format
