--- app/src/main/java/com/android3/siegertpclient/data/tournament/LeagueTournament.kt
@@ -1,10 +1,0 @@
-package com.android3.siegertpclient.data.tournament
-
-class LeagueTournament(
-    val currentState: String,
-    val gameList: List<Any>,
-    val leagueTable: Any,
-    val `open`: Boolean,
-    val participantList: List<String>,
-    val tournamentId: String
-) : Tournament2(0, TournamentDetail("","","","","","","",""), "", "")
