--- app/src/main/java/com/android3/siegertpclient/data/tournament/SpecifiedAttributes.kt
@@ -1,5 +1,0 @@
-package com.android3.siegertpclient.data.tournament
-
-data class SpecifiedAttributes(
-    val leagueTable: Any
-)
