--- app/src/main/java/com/android3/siegertpclient/data/tournament/Tournament2.kt
@@ -1,8 +1,0 @@
-package com.android3.siegertpclient.data.tournament
-
-open class Tournament2(
-    var maxParticipantNumber: Int,
-    var tournamentDetail: TournamentDetail,
-    var tournamentName: String,
-    var type: String
-)
