+++ app/src/main/java/com/android3/siegertpclient/data/tournament/TournamentDetail.kt
@@ -1,12 +1,12 @@
 package com.android3.siegertpclient.data.tournament
 
-open class TournamentDetail(
-    var adminId: String,
-    var endTime: String,
-    var location: String,
-    var participantForm: String,
-    var registrationDeadline: String,
-    var startTime: String,
-    var tournamentTypes: String,
-    var typeOfGame: String
+data class TournamentDetailXX(
+    val adminId: String,
+    val endTime: String,
+    val location: String,
+    val participantForm: String,
+    val registrationDeadline: String,
+    val startTime: String,
+    val tournamentTypes: String,
+    val typeOfGame: String
 )
