--- app/src/main/java/com/android3/siegertpclient/data/tournament/TournamentDetailXX.kt
@@ -1,12 +1,0 @@
-package com.android3.siegertpclient.data.tournament
-
-data class TournamentDetailXX(
-    val adminId: String,
-    val endTime: String,
-    val location: String,
-    val participantForm: String,
-    val registrationDeadline: String,
-    val startTime: String,
-    val tournamentTypes: String,
-    val typeOfGame: String
-)
